0.0.4

+++ YFCommonTools/Categories/NSString+Extension_YF.m
@@ -246,9 +246,10 @@ static inline NSString* StringIsEmpty (NSString *string){
         return nil;
     
     const char *value = [self UTF8String];
-    
+    // extern unsigned char *CC_MD5(const void *data, CC_LONG len, unsigned char *md)
+
     unsigned char outputBuffer[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(value, strlen(value), outputBuffer);
+    CC_MD5(value, (unsigned int)strlen(value), outputBuffer);
     
     NSMutableString *outputString = [[NSMutableString alloc] initWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(NSInteger count = 0; count < CC_MD5_DIGEST_LENGTH; count++){
